# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Building Charges & Payments API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /buildings:
    get:
      operationId: listBuildings
      tags: [buildings]
      summary: List all buildings
      responses:
        "200":
          description: List of buildings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Building"
    post:
      operationId: createBuilding
      tags: [buildings]
      summary: Create a building
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuildingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Building"
  /buildings/{id}:
    get:
      operationId: getBuilding
      tags: [buildings]
      summary: Get a building by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Building found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Building"
        "404":
          description: Not found
    patch:
      operationId: updateBuilding
      tags: [buildings]
      summary: Update a building
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuildingUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Building"
    delete:
      operationId: archiveBuilding
      tags: [buildings]
      summary: Archive a building
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Archived

  # --- UNITS ---
...
  /charges:
    get:
      operationId: listCharges
      tags: [charges]
      summary: List charges with filters
      parameters:
        - name: buildingId
          in: query
          schema:
            type: integer
        - name: unitId
          in: query
          schema:
            type: integer
        - name: personId
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
            enum: [actual, forecast]
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, paid, cancelled]
        - name: role
          in: query
          schema:
            type: string
            enum: [owner, tenant]
        - name: nameAr
          in: query
          schema:
            type: string
        - name: floor
          in: query
          schema:
            type: integer
        - name: includeArchived
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of charges
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Charge"
    post:
      operationId: createCharge
      tags: [charges]
      summary: Create a charge record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChargeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Charge"
  /charges/{id}:
    get:
      operationId: getCharge
      tags: [charges]
      summary: Get a charge by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Charge found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Charge"
        "404":
          description: Not found
    patch:
      operationId: updateCharge
      tags: [charges]
      summary: Update a charge
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChargeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Charge"
  /charges/{id}/cancel:
    post:
      operationId: cancelCharge
      tags: [charges]
      summary: Cancel a charge with reason
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelInput"
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Charge"

  # --- DASHBOARD ---
...
components:
  schemas:
...
    Building:
      type: object
      required: [id, nameAr, code]
      properties:
        id:
          type: integer
        nameAr:
          type: string
        code:
          type: string
        addressAr:
          type: ["string", "null"]
        archived:
          type: boolean
        createdAt:
          type: string
    BuildingInput:
      type: object
      required: [nameAr, code]
      properties:
        nameAr:
          type: string
        code:
          type: string
        addressAr:
          type: string
    BuildingUpdate:
      type: object
      properties:
        nameAr:
          type: string
        code:
          type: string
        addressAr:
          type: string
...
    Charge:
      type: object
      required: [id, unitId, personId, year, month, amount, type, status]
      properties:
        id:
          type: integer
        unitId:
          type: integer
        unitRef:
          type: ["string", "null"]
        buildingId:
          type: ["integer", "null"]
        buildingNameAr:
          type: ["string", "null"]
        floor:
          type: ["integer", "null"]
        personId:
          type: integer
        personNameAr:
          type: ["string", "null"]
        personRole:
          type: ["string", "null"]
        year:
          type: integer
        month:
          type: integer
        amount:
          type: number
        type:
          type: string
          enum: [actual, forecast]
        status:
          type: string
          enum: [pending, paid, cancelled]
        paidAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        cancelReason:
          type: ["string", "null"]
        archived:
          type: boolean
        createdAt:
          type: string
    ChargeInput:
      type: object
      required: [unitId, personId, year, month, amount, type]
      properties:
        unitId:
          type: integer
        personId:
          type: integer
        year:
          type: integer
        month:
          type: integer
        amount:
          type: number
        type:
          type: string
          enum: [actual, forecast]
        status:
          type: string
          enum: [pending, paid, cancelled]
        paidAt:
          type: string
        notes:
          type: string
    ChargeUpdate:
      type: object
      properties:
        amount:
          type: number
        status:
          type: string
          enum: [pending, paid, cancelled]
        paidAt:
          type: string
        notes:
          type: string
        type:
          type: string
          enum: [actual, forecast]
    CancelInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
...